Alterações: ESCL018/ESCL018B/ESCL027/ESCL034
-ESCL018/ESCL018B aceitar quantidade decimal no inventario.
-ESCL027 solicitar o número de série para o reparo.
-ESCL034 não permitir a leitura do mesmo Item X Serie na mesma sessao.

+++ CollectorQi/Models/ESCL018/Inventario.cs
@@ -7,7 +7,7 @@
         public string CodDepos { get; set; }
         public string Localizacao { get; set; }
         public string Lote { get; set; }
-        public int QuantidadeDigitada { get; set; }
+        public decimal QuantidadeDigitada { get; set; }
         public string CodigoBarras { get; set; }
     }
 }
